+++ src/entrypoints/abi/distance_matrix.rs
@@ -2,5 +2,5 @@
 pub struct ABIDistanceMatrixEntry {
     pub from: u32,
     pub to: u32,
-    pub distance: f64,
+    pub distance: f32,
 }
